New LICENSE file (CC BY SA 4.0)
Didn't have an option to start the GH repo with this, did a cut and paste myself from creativecommons.org.
Caveat is that the format of this file likely will need a changeover, as I feel there is a better way to represent the CC LICENSE file but loking at e.g. Fedora (https://docs.fedoraproject.org/en-US/fedora/f30/install-guide/ ) only gives a clue as that is an entire disclaimer. We may want a disclaimer in a finished/printed/rendered version, but not necessarily all of that in the source code LICENSE file. But it may be there is some nuance to how the CC license is best called. Thus we may end up adjusting the _contents_and_format_ of this file but not relicensing in doing so.