Updating to reflect new repo name
Details have been evolving, this is one of those changes.

In discussions with others, I agreed with their point about compendium being a bit obscure and maybe open to interpretation not just in meaning but ... friendliness? Whereas a guidebook guides you through things, and that's what we're about.

When I wrote this description originally I was envisioning a scope that might cover recommendations across the board for how we conduct things in an open source project -- infrastructure, source repositories, QA/testing & CI/D, and so forth. However, this guidebook is truly focused on community management best practices. We'll see what comes next, next.