Removing 3r3/4th level topics to tighten ToC
This tightens the outline to just cover the sections and chapters.

The third and beyond levels that I just removed represent various topics that MAY appear in a chapter, depending on what the final decisions of the authors of that chapter--they were topics the original outline creators thought would be relevant, with an eye on the 1.0 source material. We are sure that other topics will arise, so in removing them here from the outline we are reducing the wearying effect on (new) contributors in seeing what sees like a very large and very complete outline at first glance.

Instead, the topics are now covered as an issue-per-topic, and then are assigned one or more labels to show they belong in one or more chapters. We may want to have a chapter-header of some sort that tracks the topics within a chapter as they are being worked on?