Adding specific mentorship content, rearranging c
I had been thinking that the "From users to contributors" content would cover mentoring. This current version doesn't really, and in thinking about it, there is a wider idea here.

Rather than just creating a mentorship program and checkbox, I think our opinion is the best practice is create a culture of mentorship in the community. So, yes, a formal program (which helps especially those who like formal programs), but to identify the qualities of mentorship and imbue them across the community.

Thus it's more than just "get someone from a user to a contributor, all done", it's recognizing that we all mentor each other, and always need mentoring, toward the reality of continuous improvement as the antidote for stasis.