Rolling in author approved changes from PR
Leslie replied to the merged PR here:

https://github.com/theopensourceway/guidebook/pull/151/files/bb8cafc112d38f186329d95d649a45709ebbe21c

I went through each of those changes and resolved them in this version rather than waiting for the future. Leslie left clear intent for each of the suggestions from Paula. Those suggestions were equally crystal clear. So I boldly made the changes for this version, which I think will help that extra % of readers who benefit from the various formatting considerations, etc.

We can move forward with this chapter for publishing in the preview release, and Leslie can still revisit each of these decision points to improve, reverse, rework, or reimagine the flavor and flow.