Adding modified toolchain bits
This is a minor evolution of the build tools. It does the following:

1. Copies files listed in 'chaplist' to temp build files.
2. Adjusts the nesting of sections in the temp build files.
3. Builds an HTML single-page book usimg the temp build files.
4. Cleans up the temp build files.

Running 'book-build-0.2.sh' should produce an attractive HTML output
without tweaking with the original files. This way the content files
can be in whatever flux with open PRs, etc. and the build tools are
non-destructive to the original files in this way.