syntax standards formatting fixes
Two things happening here:
. One sentence per line, which makes copy editing easier. But is harder to maintain when in the writing phase, so it has been normal to ignore that until the author is ready for development or copy edit, which this seems to be.
. Using the 'term' Asciidoc format for lists of extended topics within a sub-section, which is what Ray was doing here with formatting. We are piggybacking a tiny bit on the semantic meaning of a "term" but in general it works, and it's consistent across the guide.

I think we can merge at this point to move forward. Unless Ray has additional material to add to this branch right now. None of my edits so far are hard to recreate or duplicate if a revert is needed to avoid a nasty merge conflict.